Removed the exe.py script. Updated the README.md file. Added the requirements.txt file. Updated the .gitignore file

+++ scripts/exe.py
@@ -1,19 +1,27 @@
-"""Generate executable files for the server and client scripts using pyinstaller"""
-
 import os
 from shutil import rmtree
 
 
+REMOVE_BUILD = True
+REMOVE_SPEC = True
+
+
 scripts = ("server.py", "client.py")
+
 
 for script in scripts:
     try: 
         print(f'Generating executable for {script}...')
-        os.system(f'python -m PyInstaller {script} --onefile --distpath ../apps')
-        
-        print("Removing build files...")
-        os.remove(f'{script[:-3]}.spec')
-        rmtree('build')
+        os.system(f'python -m PyInstaller {script} --onefile --distpath /apps')
+
+        if REMOVE_SPEC:
+            print(f'Removing {script[:-3]}.spec file...')
+            os.remove(f'{script[:-3]}.spec')
+
+        if REMOVE_BUILD:
+            print('Removing build directory...')
+            rmtree('build')
+
     except Exception as e:
         print(f'Error generating executable for {script}: {e}')
